requirements.txt
streamlit
crawl4ai
crewai
crewai_tools
google-generativeai
Markdown
python-docx
beautifulsoup4
playwright
selenium
pandas
openpyxl
requests
